refactor(webhook_service/list_webhooks): extract the function (SCRUM-157)
