+++ src/webhook_service/list_webhooks/lambda_function.py
@@ -6,10 +6,10 @@ from typing import Dict
 
 import boto3
 from boto3.dynamodb.conditions import Key
+from get_total_count import get_total_count
 
 dynamodb = boto3.resource("dynamodb")
 table = dynamodb.Table(os.getenv("DYNAMODB_TABLE"))
-total_count_table = dynamodb.Table(os.getenv("DYNAMODB_TABLE_TOTAL_COUNT"))
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -21,17 +21,6 @@ class DecimalEncoder(json.JSONEncoder):
             return float(obj)
         return super(DecimalEncoder, self).default(obj)
 
-def get_total_count(webhook_type: str) -> int:
-    """
-    Get the total count of items for a specific webhook_type.
-    """
-    try:
-        response = total_count_table.get_item(Key={"webhook_type": webhook_type})
-        total_count = response.get("Item", {}).get("total_count", 0)
-        return int(total_count)
-    except Exception as e:
-        logger.error(f"Error fetching total count for {webhook_type}: {str(e)}")
-        raise
 
 def lambda_handler(event: Dict, context) -> Dict:
     """
@@ -68,7 +57,6 @@ def lambda_handler(event: Dict, context) -> Dict:
                 "headers": {"Content-Type": "application/json"},
                 "body": json.dumps({"error": "Invalid limit value."}),
             }
-        
         
         
         elif (page - 1) * limit >= total_count:
